src/test/java/ru/yandex/practicum/filmorate/models/FilmControllerTest.java
package ru.yandex.practicum.filmorate.models;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import ru.yandex.practicum.filmorate.controllers.FilmController;

import java.time.LocalDate;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

public class FilmControllerTest {

    Film film1 = new Film(1, "N1", "D1",
            LocalDate.of(2024, 1, 1), 100);

    private FilmController filmController;

    @BeforeEach
    void setUp() {
        filmController = new FilmController();
    }

    @Test
    @DisplayName(value = "Тест GET запроса на получение всех фильмов")
    public void test_1_getFilms() {
        ResponseEntity<List<Film>> response = filmController.getFilms();
        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertNotNull(response.getBody());
        assertTrue(response.getBody().isEmpty());
    }
}